test(e2e): share OpenClaw output parser
